Database created, basic functionality implemented
Database in place, takesinfo from submit page for now, till we get a crawler working.

Also created a PlayRandom method in the WebVid controller that returns a random video from the database.

Need to figure out how to make it so it doesnt return the same thing twice in a row, or twice in the same session maybe?