Installed react-player. added a ReactPlayerTest page.
It works as intended. As long as we get the hard urls from PinkBike vids (they dont play right now) we can start building some controls and making things look pretty.

Also it's all react embedded in html ATM. so we'll have to pull out what we need and make it local.